+++ tests/test_save_multiple.py
@@ -6,7 +6,7 @@ from tornado.web import HTTPError
 import jupytext
 from jupytext.contentsmanager import TextFileContentsManager
 from jupytext.compare import compare_notebooks
-from .utils import list_all_notebooks
+from .utils import list_all_notebooks, list_py_notebooks
 
 jupytext.file_format_version.FILE_FORMAT_VERSION = {}
 
@@ -50,7 +50,7 @@ def test_ipynb_is_ok(nb_file, tmpdir):
     compare_notebooks(nb, nb2)
 
 
-@pytest.mark.parametrize('nb_file', list_all_notebooks('.ipynb'))
+@pytest.mark.parametrize('nb_file', list_py_notebooks('.ipynb'))
 def test_all_files_created(nb_file, tmpdir):
     nb = jupytext.readf(nb_file)
     tmp_ipynb = 'notebook.ipynb'